workspace/scratch_project/requirements.txt
biopython
rdkit-pypi
pandas
scikit-learn
